datagen fully working, moved some funcs from rgeo
Also removed New() and rgeo.ReverseGeocode since the rgeo already
exists. In its current state it is significantly slower than
rgeo.ReverseGeocode but faster than the old ReverseGeocode, this is
because converting from geom geometries to s2 polygons seems to be most
of the work. I could write a new New() function to do that ahead of time
but I'd rather just figure out a new way to check if a point is within a
geometry without using s2 at all.