Migrate base64 go files to go:embed
This mostly improves compile times and the ease of understanding.

I could not measure an improvment in unpacking, even though the
base64 decoding step got removed from the process.

With this I guess we can conclude that go:embed operations are
not super fast. Since base64 decoding is normally quite high cost.

Closes https://github.com/sams96/rgeo/issues/15